shift fc data on utils
